+++ Sources/XcodeSupport/XcodeTarget.swift
@@ -2,7 +2,6 @@ import Foundation
 import XcodeProj
 import SystemPackage
 import PeripheryKit
-import Shared
 
 final class XcodeTarget {
     let project: XcodeProject